:sparkles: add portals

+++ lib/ecs/components/movable.dart
@@ -3,8 +3,8 @@ mixin MovableComponent {
 }
 
 class Speed {
-  double dx;
-  double dy;
+  int dx;
+  int dy;
 
   Speed({this.dx = 0, this.dy = 0})
       : assert(dx != null),
